Platform works

+++ src/levels/level.rs
@@ -71,9 +71,9 @@ pub struct ThingConfig {
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct PlatformConfig {
-    pub x: f32,
-    pub y: f32,
-    pub width: f32,
-    pub height: f32,
+    pub x: i32,
+    pub y: i32,
+    pub width: i32,
+    pub height: i32,
     pub color_flag: ColorFlag,
 }
